+++ ports/freertos/include/memfault/ports/freertos.h
@@ -1,0 +1,9 @@
+#pragma once
+
+//! @file
+//!
+//! Copyright (c) Memfault, Inc.
+//! See License.txt for details
+
+//! Should be called prior to making any Memfault SDK calls
+void memfault_freertos_port_boot(void);